feat: Addresses
Added new domain classes "Address" and "PartyAddress". This modelling is to keep the idea of an address and the idea of a Party logically separate. It also will allow for many-many relationships down the line, and also for addresses to sit both on a party, but also an institution etc. While this may not be necessary for OA per se, it will be necessary in other projects. Since we hope to migrate this work over to using something more general in future, this structure should hopefully allow for a more seamless transition.

MODOA-18