feat: Tweo new fields for books
Added two fields: `bookDateOfPublication` and `bookPlaceOfPublication` to the publicatin request. In future we may wish to expand these to proper title instances, but we don't currently have enough citation data to stand those up anyway, so store here as is for now with a view to maybe migrating later on.